Added aravis camera callback functionality
Before: the camera started writing images to a stream slowly filling up
the buffers. In the task thread a function would read out those images
from the buffers at the same frequency as the camera wrote them.

Now, we can also set the camera to emit signals when a buffer is done
and that signal triggers the plugin to read the buffer out.

Both options are available under the new flag aravis_callback. If set to
true, it uses the second option, false for the first. Default is true.

Additionally fixed a few bugs made through spelling mistakes and now the
image size in pixels is available through status.