Added comment for two function
ip and port callback require a command afterwards so the CLI specifies
this now.

That's because it's 'intended' behavior: you want to be able to change
the ip and then run a command, but if you use typer.exit which is how
you tell typer not to wait for any more commands it exits without it.

I'll try to find a better solution